+++ src/main/java/wooteco/subway/auth/config/AuthenticationPrincipalArgumentResolver.java
@@ -5,23 +5,17 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
-import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.method.support.ModelAndViewContainer;
-import wooteco.subway.auth.application.AuthService;
 import wooteco.subway.auth.domain.AuthenticationPrincipal;
 import wooteco.subway.auth.infrastructure.AuthorizationExtractor;
-import wooteco.subway.auth.infrastructure.JwtTokenProvider;
 import wooteco.subway.member.application.MemberService;
-import wooteco.subway.member.domain.Member;
 import wooteco.subway.member.dto.MemberResponse;
 
 public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArgumentResolver {
-    private AuthService authService;
     private MemberService memberService;
 
-    public AuthenticationPrincipalArgumentResolver(AuthService authService, MemberService memberService) {
-        this.authService = authService;
+    public AuthenticationPrincipalArgumentResolver(MemberService memberService) {
         this.memberService = memberService;
     }
 
@@ -34,7 +28,7 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
     public Object resolveArgument(MethodParameter parameter, ModelAndViewContainer mavContainer, NativeWebRequest webRequest, WebDataBinderFactory binderFactory) {
         String token = AuthorizationExtractor.extract(
             Objects.requireNonNull(webRequest.getNativeRequest(HttpServletRequest.class)));
-        String email = authService.getPayLoad(token);
+        String email = memberService.getPayLoad(token);
         MemberResponse memberResponse = memberService.findMemberByEmail(email);
         return memberResponse;
     }
