+++ src/main/java/wooteco/subway/auth/config/AuthenticationPrincipalConfig.java
@@ -3,18 +3,15 @@ package wooteco.subway.auth.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import wooteco.subway.auth.application.AuthService;
 
 import java.util.List;
 import wooteco.subway.member.application.MemberService;
 
 @Configuration
 public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
-    private final AuthService authService;
     private final MemberService memberService;
 
-    public AuthenticationPrincipalConfig(AuthService authService, MemberService memberService) {
-        this.authService = authService;
+    public AuthenticationPrincipalConfig(MemberService memberService) {
         this.memberService = memberService;
     }
 
@@ -25,6 +22,6 @@ public class AuthenticationPrincipalConfig implements WebMvcConfigurer {
 
     @Bean
     public AuthenticationPrincipalArgumentResolver createAuthenticationPrincipalArgumentResolver() {
-        return new AuthenticationPrincipalArgumentResolver(authService, memberService);
+        return new AuthenticationPrincipalArgumentResolver(memberService);
     }
 }
