+++ src/main/java/wooteco/subway/member/application/MemberService.java
@@ -1,6 +1,11 @@
 package wooteco.subway.member.application;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.auth.dto.TokenRequest;
+import wooteco.subway.auth.dto.TokenResponse;
+import wooteco.subway.auth.infrastructure.JwtTokenProvider;
+import wooteco.subway.exception.auth.LoginFailEmailException;
+import wooteco.subway.exception.auth.LoginWrongPasswordException;
 import wooteco.subway.exception.member.NotRegisteredMemberException;
 import wooteco.subway.member.dao.MemberDao;
 import wooteco.subway.member.domain.Member;
@@ -9,9 +14,12 @@ import wooteco.subway.member.dto.MemberResponse;
 
 @Service
 public class MemberService {
-    private MemberDao memberDao;
 
-    public MemberService(MemberDao memberDao) {
+    private final JwtTokenProvider tokenProvider;
+    private final MemberDao memberDao;
+
+    public MemberService(JwtTokenProvider tokenProvider, MemberDao memberDao) {
+        this.tokenProvider = tokenProvider;
         this.memberDao = memberDao;
     }
 
@@ -36,5 +44,24 @@ public class MemberService {
 
     public void deleteMember(Long id) {
         memberDao.deleteById(id);
+    }
+
+    public TokenResponse createToken(TokenRequest tokenRequest) {
+        String email = tokenRequest.getEmail();
+        String accessToken = tokenProvider.createToken(email);
+        return new TokenResponse(accessToken);
+    }
+
+    private void checkAvailableLogin(TokenRequest tokenRequest) {
+        String email = tokenRequest.getEmail();
+        String password = tokenRequest.getPassword();
+        Member member = memberDao.findByEmail(email).orElseThrow(LoginFailEmailException::new);
+        if (!member.isSamePassword(password)) {
+            throw new LoginWrongPasswordException();
+        }
+    }
+
+    public String getPayLoad(String tokenName) {
+        return tokenProvider.getPayload(tokenName);
     }
 }
